+++ src/protocol/attribution/credit_capping.rs
@@ -1,10 +1,10 @@
 use super::{if_else, CreditCappingInputRow, CreditCappingOutputRow, InteractionPatternStep};
 use crate::error::Error;
 use crate::ff::Field;
+use crate::protocol::basics::mul::SecureMul;
 use crate::protocol::boolean::random_bits_generator::RandomBitsGenerator;
 use crate::protocol::boolean::{local_secret_shared_bits, BitDecomposition, BitwiseLessThan};
 use crate::protocol::context::{Context, SemiHonestContext};
-use crate::protocol::mul::SecureMul;
 use crate::protocol::{RecordId, Substep};
 use crate::secret_sharing::Replicated;
 use futures::future::{try_join, try_join_all};
