+++ src/protocol/basics/mul/semi_honest.rs
@@ -210,7 +210,7 @@ pub async fn multiply_one_share_mostly_zeroes<F: Field>(
 #[cfg(all(test, not(feature = "shuttle")))]
 mod regular_mul_tests {
     use crate::ff::{Field, Fp31};
-    use crate::protocol::mul::SecureMul;
+    use crate::protocol::basics::mul::SecureMul;
     use crate::protocol::{QueryId, RecordId};
     use crate::rand::{thread_rng, Rng};
     use crate::test_fixture::{Reconstruct, Runner, TestWorld};
@@ -303,7 +303,7 @@ mod specialized_mul_tests {
 
     use super::{multiply_one_share_mostly_zeroes, multiply_two_shares_mostly_zeroes};
     use crate::ff::Fp31;
-    use crate::protocol::mul::test::{SpecializedA, SpecializedB, SpecializedC};
+    use crate::protocol::basics::mul::test::{SpecializedA, SpecializedB, SpecializedC};
     use crate::protocol::{QueryId, RecordId};
     use crate::rand::{thread_rng, Rng};
     use crate::test_fixture::{Reconstruct, Runner, TestWorld};
